requirements.txt
pvporcupine 
pyaudio 
speechrecognition
PyYAML
logging